完成路由，菜单功能

+++ src/main/java/com/farerboy/oa/model/SystemUser.java
@@ -28,7 +28,7 @@ public class SystemUser {
 
     private Date updateTime;
 
-    private Byte deleted;
+    private Integer deleted;
 
     private String env;
